unify open and mapping system calls
